# packages/protocol/openapi/agentnet.openapi.yaml
openapi: 3.1.0
info:
  title: AgentNet API
  description: AgentNet is a centralized Agent Relay Platform for registering agents,
    routing asynchronous tasks, tracking delivery state, and enforcing cross-agent
    approval policy. WebSocket agent transport is available at `/v1/ws` and documented
    in `docs/openapi.md`.
  version: 0.1.0
paths:
...
  /v1/dashboard/admin/network/scopes:
    get:
      tags:
      - dashboard-admin
      summary: List Network Scopes Admin
      description: List all network scopes across the platform. Admin-only.
      operationId: list_network_scopes_admin_v1_dashboard_admin_network_scopes_get
      parameters:
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: scope_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Scope Type
      - name: agentnet_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkScopeListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyBearer: []
      - LegacyApiKeyHeader: []
  /v1/dashboard/admin/network/scopes/{scope_id}:
    get:
      tags:
      - dashboard-admin
      summary: Get Network Scope Detail
      description: Get a single network scope by ID. Admin-only.
      operationId: get_network_scope_detail_v1_dashboard_admin_network_scopes__scope_id__get
      parameters:
      - name: scope_id
        in: path
        required: true
        schema:
          type: string
          title: Scope Id
      - name: agentnet_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkScopeDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyBearer: []
      - LegacyApiKeyHeader: []
  /v1/dashboard/admin/network/zones:
    get:
      tags:
      - dashboard-admin
      summary: List Network Zones Admin
      description: List all network zones across the platform. Admin-only.
      operationId: list_network_zones_admin_v1_dashboard_admin_network_zones_get
      parameters:
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: zone_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone Type
      - name: scope_id_filter
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Scope Id Filter
      - name: agentnet_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkZoneListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyBearer: []
      - LegacyApiKeyHeader: []
  /v1/dashboard/admin/network/zones/{zone_id}:
    get:
      tags:
      - dashboard-admin
      summary: Get Network Zone Detail
      description: Get a single network zone by ID. Admin-only.
      operationId: get_network_zone_detail_v1_dashboard_admin_network_zones__zone_id__get
      parameters:
      - name: zone_id
        in: path
        required: true
        schema:
          type: string
          title: Zone Id
      - name: agentnet_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkZoneDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyBearer: []
      - LegacyApiKeyHeader: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NetworkScopeDetailResponse:
      properties:
        scope_id:
          type: string
          title: Scope Id
        scope_name:
          type: string
          title: Scope Name
        scope_type:
          type: string
          title: Scope Type
        user_id:
          type: string
          title: User Id
        username:
          type: string
          title: Username
        network_cidr:
          anyOf:
          - type: string
          - type: 'null'
          title: Network Cidr
        agent_ids:
          items:
            type: string
          type: array
          title: Agent Ids
        zone_ids:
          items:
            type: string
          type: array
          title: Zone Ids
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - scope_id
      - scope_name
      - scope_type
      - user_id
      - username
      - created_at
      - updated_at
      title: NetworkScopeDetailResponse
    NetworkScopeListItem:
      properties:
        scope_id:
          type: string
          title: Scope Id
        scope_name:
          type: string
          title: Scope Name
        scope_type:
          type: string
          title: Scope Type
        user_id:
          type: string
          title: User Id
        username:
          type: string
          title: Username
        network_cidr:
          anyOf:
          - type: string
          - type: 'null'
          title: Network Cidr
        agent_count:
          type: integer
          title: Agent Count
          default: 0
        zone_count:
          type: integer
          title: Zone Count
          default: 0
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - scope_id
      - scope_name
      - scope_type
      - user_id
      - username
      - created_at
      - updated_at
      title: NetworkScopeListItem
    NetworkScopeListResponse:
      properties:
        scopes:
          items:
            $ref: '#/components/schemas/NetworkScopeListItem'
          type: array
          title: Scopes
        total:
          type: integer
          title: Total
        offset:
          type: integer
          title: Offset
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - scopes
      - total
      - offset
      - limit
      title: NetworkScopeListResponse
    NetworkZoneDetailResponse:
      properties:
        zone_id:
          type: string
          title: Zone Id
        zone_name:
          type: string
          title: Zone Name
        zone_type:
          type: string
          title: Zone Type
        scope_id:
          type: string
          title: Scope Id
        scope_name:
          type: string
          title: Scope Name
        parent_zone_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Zone Id
        relay_node_ids:
          items:
            type: string
          type: array
          title: Relay Node Ids
        zone_metadata:
          type: object
          title: Zone Metadata
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - zone_id
      - zone_name
      - zone_type
      - scope_id
      - scope_name
      - created_at
      - updated_at
      title: NetworkZoneDetailResponse
    NetworkZoneListItem:
      properties:
        zone_id:
          type: string
          title: Zone Id
        zone_name:
          type: string
          title: Zone Name
        zone_type:
          type: string
          title: Zone Type
        scope_id:
          type: string
          title: Scope Id
        scope_name:
          type: string
          title: Scope Name
        parent_zone_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Zone Id
        relay_node_count:
          type: integer
          title: Relay Node Count
          default: 0
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - zone_id
      - zone_name
      - zone_type
      - scope_id
      - scope_name
      - created_at
      - updated_at
      title: NetworkZoneListItem
    NetworkZoneListResponse:
      properties:
        zones:
          items:
            $ref: '#/components/schemas/NetworkZoneListItem'
          type: array
          title: Zones
        total:
          type: integer
          title: Total
        offset:
          type: integer
          title: Offset
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - zones
      - total
      - offset
      - limit
      title: NetworkZoneListResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    ApiKeyBearer:
      type: http
      scheme: bearer
      description: 'REST API key in `Authorization: Bearer ak_...`.'
    LegacyApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: Legacy REST API key header.
...
servers:
- url: http://localhost:8000
  description: Local development